Moved ChartController.ChartRange and ChartGranularity in a separate file.
