+++ src/charts/core/ChartController.ts
@@ -22,6 +22,7 @@ import {computed, Ref, ref, watch, WatchStopHandle} from "vue";
 import {Chart, ChartConfiguration} from 'chart.js/auto';
 import {ThemeController} from "@/components/ThemeController.ts";
 import {RouteManager} from "@/utils/RouteManager.ts";
+import {ChartRange} from "@/charts/core/ChartRange.ts";
 
 export enum ChartState {
     unsupported,
@@ -192,64 +193,6 @@ export abstract class ChartController<M> {
     }
 }
 
-export enum ChartRange {            // Matching granularity
-    day = "day",                    // => hour
-    year = "year",                  // => month
-    all = "all"                     // => year
-}
-
-export enum ChartGranularity {
-    hour = "hour",
-    day = "day",
-    month = "month",
-    year = "year"
-}
-
 export class LoadedData<M> {
     constructor(public readonly metrics: M[], public readonly latestMetric: M|null) {}
 }
-
-
-export function computeStartDateForRange(period: ChartRange): string {
-    let result: string
-    const now = new Date()
-    switch(period) {
-        case ChartRange.all: {
-            const d = new Date(0)
-            result = d.toISOString()
-            break
-        }
-        case ChartRange.year: {
-            const y = now.getFullYear()
-            const m = now.getMonth()
-            const d = new Date(y-1, m)
-            result = d.toISOString()
-            break
-        }
-        case ChartRange.day: {
-            const d = new Date(now.getTime() - 24 * 3600 * 1000)
-            result = d.toISOString()
-            break
-        }
-    }
-    return result
-}
-
-
-export function computeGranularityForRange(period: ChartRange): ChartGranularity {
-    let result: ChartGranularity
-    switch(period) {
-        default:
-        case ChartRange.all:
-            result = ChartGranularity.month
-            break
-        case ChartRange.year:
-            result = ChartGranularity.month
-            break
-        case ChartRange.day:
-            result = ChartGranularity.hour
-            break
-    }
-    return result
-}
-
