+++ src/charts/hgraph/NetworkFeeController.ts
@@ -20,7 +20,7 @@
 
 import {GenericMetricController} from "@/charts/hgraph/GenericMetricController.ts";
 import {EcosystemMetric} from "@/charts/hgraph/EcosystemMetric.ts";
-import {ChartRange} from "@/charts/core/ChartController.ts";
+import {ChartRange} from "@/charts/core/ChartRange.ts";
 import {ThemeController} from "@/components/ThemeController.ts";
 import {RouteManager} from "@/utils/RouteManager.ts";
 import {ChartConfiguration} from "chart.js/auto";
